Fix handling sub change while serving subgroup
Before this change, a publisher (or relay) would only check the subgroup
to be served against the subscribe filter at the beginning of the subgroup.
This behavior caused problems when a `SUBSCRIBE_UPDATE` was received in the
middle of the subgroup, as the publisher kept sending subsequent objects from
the subgroup without consulting the refreshed subscribe filter which could
instruct the publisher to restrain from sending the rest of the subgroup.