BREAKING CHANGE: Retain filter type compatibility with draft-07
This can be achieved by handling LatestObject and LatestGroup as the same
filter. However, doing so disregards a requirement stated in the IETF drafts.

In particular, the draft prescribes that receiving an unknown filter type MUST
be treated as error (and since LatestGroup no longer exists, it is considered
to be like that):

> A filter type other than the above MUST be treated as error.